Ignore data race for these tests
